requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
astroid==3.3.5
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
dill==0.3.9
distro==1.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
isort==5.13.2
jiter==0.8.0
mccabe==0.7.0
openai==1.55.1
platformdirs==4.3.6
pydantic==2.10.2
pydantic_core==2.27.1
pylint==3.3.2
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
tomlkit==0.13.2
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
